Other Page

This is something different

PathIntegrationMethod.PathIntegration — Type

PathIntegration(sde, method, ts, axes...; kwargs...)

Compute a PathIntegration object for computing the response probability density function evolution of the stochastic dynamical system defined by sde.

Arguments

sde::AbstractSDE{d,k,m}: $d$-dimensional dynamic system which is subjected to an $m$-dimensional Wiener process acting on the $k...d$ coordinates.

method::DiscreteTimeSteppingMethod: method used for the time-discretisation of the dynamical system described by sde.

ts::Union{Number,AbstractArray{<:Number}}: If AbstractArray{<:Number} is provided then the stepMX-s between the times in ts are computed. If Number is provided, then it computes a single stepMX is computed between 0 and ts.

axes::Vararg{aT,d} where aT<:GridAxis: The axes spanning the space where the transitional PDF is computed for the dynamical system sde.

Keyword Arguments

discreteintegrator = d==k ? QuadGKIntegrator() : ClenshawCurtisIntegrator(): Discrete integrator to evaluate the Chapman-Kolmogorov equation

di_N = 21: Resolution of the discrete integrator. Can be an NTuple{d-k+1,<:Integer} or an AbstractArray{<:Integer} to define resolution to each integration variable

smart_integration = true: Only integrate where the transitional PDF has nonzero elements. It is approximated with the step function. Use false if (time step * diffusion) results in a wide TPDF. Usually true is the better choice.

int_limit_thickness_multiplier = 6: The "thickness" scaling of the TPDF during smart integration.

initialise_pdf = true: Initialize the response probability density function. If false, then the RPDF is initialzed as p(x) ≡ 0.

f_init = nothing: Initial RPDF as a function. If f_init is a Nothing and initialise_pdf = true then a diagonal Gaussian distribution is used as initial distribution

μ_init = nothing: Mean of the initial Gaussian distribution.

Nothing: Uses the middle of the range defined by the axis

Number: Uses the single value μ_init for each axis direction

Union{NTuple{d,<:Number},AbstractVector{<:Number}}: Individual means for each axis.

σ_init = nothing: Standard deviation of the initial diagonal Gaussian distribution.

Nothing: Uses the 1/12th of the range width in each direction defined by the axes.

Number: Uses the single value σ_init for each axis direction

Union{NTuple{d,<:Number},AbstractVector{<:Number}}: Individual standard deviations for each axis.

pre_compute = true: Compute the stepMX. This should be left unchanged if the RPDF computation is the goal.

For methods, discrete integrators, interpolators, and examples please refer to the documentation.